feat: защита веб-адреса, кода в HTML и разряды числа в фазе Bind
Веб-адрес — машинный идентификатор, а не текст. Границы адреса вынесены в
Internal/Url и стали общими для фаз Scan, Bind и автоссылки фазы Layout: раньше
каждая держала свою копию, и кавычка, завершившая адрес для одной, уезжала в href
у другой. Внутри адреса не работают правила Scan и Bind — точка в «?v=1.5» не
разделитель, «/2020-2026» не тире, одиннадцать цифр в пути не телефон, — и слово
за адресом к его хвосту не привязывается.

Автоссылка: амперсанд, начинающий сущность, больше не экранируется повторно
(«&amp;» не превращается в «&amp;amp;»), точка с запятой сущности не отрезается
как хвостовая пунктуация, а прямой апостроф отрезается только непарный — «?q='x y'»
законен по RFC 3986.

Markdig: содержимое code, pre, kbd, samp, script, style и textarea, записанных
прямо в Markdown, защищено — для Markdig это отдельные узлы, и без учёта таких зон
типограф правил код и клавиши.

digitGrouping теперь включает и проход фазы Bind: разбить «1000000» умеет Scan, а
сделать неразрывными уже стоящие в «1 000 000» пробелы — только Bind, где есть токены.

Строчный тег для правил прозрачен: «25 <b>°C</b>» обрабатывается как «25 °C».

+++ src/Typographer/Internal/CharBuffer.cs
@@ -58,7 +58,9 @@ internal struct CharBuffer
 
     private void EnsureCapacity(int additional)
     {
-        int required = _length + additional;
+        // Требуемая длина тоже считается в long: у буфера, доросшего до предела массива,
+        // сумма в int уходит в минус и проходит проверку как «места хватает».
+        long required = (long)_length + additional;
         if (required <= _array.Length)
         {
             return;
